requirements.txt
# Deep Learning Framework
torch>=1.9.0
torchvision>=0.10.0

# Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Image Processing
Pillow>=8.3.0
opencv-python>=4.5.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.0.0

# Web App Framework
streamlit>=1.10.0

# Additional utilities
tqdm>=4.62.0
